PostgreSQL serving layer, end-to-end and verified
The schema and loader existed but had never run against a live server. Running
them found two bugs, and the verification step that found them is now permanent.

Live cluster
- PostgreSQL 17.2 from the binaries-only distribution: no installer, no admin
  rights, no Windows service. localhost:5433, scram-sha-256 over TCP.
- scripts/pg.ps1 manages setup/start/stop/status/psql; docs/POSTGRES_SETUP.md
  documents it. A generated password is written to .env (gitignored), never
  committed.
- Loaded and verified: 4,934,699 sessions, 66,386 products, 57,483 performance
  rows. Both CHECK constraints held.

Bug 1 - HUGEINT cannot round-trip through pandas
- DuckDB's sum() returns HUGEINT (int128); pandas has no int128 dtype, so the
  column arrived as float64 and to_csv wrote "0.0", which PostgreSQL rejects
  for an INTEGER column.
- Fixed with ::BIGINT casts in 02_marts.sql and 03_ml_features.sql, plus a
  guard in the loader that names the offending column instead of failing deep
  inside COPY with a row number.

Bug 2 - timestamps were machine-dependent
- to_timestamp() returns TIMESTAMP WITH TIME ZONE, and hour()/dayofweek()
  extract in the SESSION's time zone. hour_of_day was therefore a property of
  the machine running the pipeline, not of the data.
- Measured: switching to epoch_ms() (naive UTC) changed hour_of_day for 100%
  of the 26.4M staged events and day_of_week for 34.6% of them.
- Caught by verify_postgres.py, which flagged the two timestamp columns as
  disagreeing between engines. Nothing else would have surfaced it.
- Two tests now recompute both columns in Python from the raw epoch ms, so the
  guard cannot be satisfied by the same DuckDB expression it is checking.

Load tuning: 518.5s -> 156.7s for fct_session, via one wrong turn worth
recording. Profiling blamed pandas (57.9%), so pandas was removed -- and the
time barely moved (507.1s). The profile had copied into a TEMP table, which has
no indexes, making COPY look free; on the real table, maintaining a 587 MB
primary-key index over a 64-char hash was the actual cost. Dropping and
rebuilding indexes around the load is what worked. CHECK constraints are
deliberately kept during load.

Also: hourly_pattern.png plus an intra-day section. The UTC curve is a textbook
retail day -- trough 07:00 UTC, peak 00:00 UTC -- which places the retailer near
UTC-5. That is an inference from shape, not a disclosed fact, and nothing
depends on it; it is stated because after a timezone bug, evidence that the
conversion behaves sensibly is worth showing.

+++ sql/duckdb/02_marts.sql
@@ -70,13 +70,20 @@ WITH ev AS (
     GROUP BY session_id_hash
 ),
 srch AS (
+    -- The ::BIGINT casts are load-bearing, not cosmetic. DuckDB's sum() on an
+    -- integer column returns HUGEINT (int128) to be overflow-safe. pandas has
+    -- no int128, so a HUGEINT column silently arrives as float64, and
+    -- to_csv then writes "0.0" -- which PostgreSQL rejects for an INTEGER
+    -- column. These counts are small by construction, so the width is pure
+    -- overhead; casting here fixes the type at the source rather than papering
+    -- over it in the loader.
     SELECT
         session_id_hash,
         count(*)                                        AS n_searches,
         count(*) FILTER (WHERE is_zero_result)          AS n_zero_result_searches,
-        sum(n_clicks)                                   AS n_search_clicks,
-        sum(len(valid_clicked_skus))                    AS n_valid_search_clicks,
-        sum(n_results)                                  AS n_search_impressions
+        sum(n_clicks)::BIGINT                           AS n_search_clicks,
+        sum(len(valid_clicked_skus))::BIGINT            AS n_valid_search_clicks,
+        sum(n_results)::BIGINT                          AS n_search_impressions
     FROM stg_search
     GROUP BY session_id_hash
 )
